fix : move setMessagesByDate into chatMessageList
